feat(web): Broke up Satisfaction section into it's own subcomponents
