+++ web/src/components.d.ts
@@ -26,7 +26,9 @@ declare module 'vue' {
     PlannerFactoryList: typeof import('./components/planner/PlannerFactoryList.vue')['default']
     PlannerFactoryProducts: typeof import('./components/planner/PlannerFactoryProducts.vue')['default']
     PlannerFactorySatisfaction: typeof import('./components/planner/PlannerFactorySatisfaction.vue')['default']
+    PlannerFactorySatisfactionBuildings: typeof import('./components/planner/PlannerFactorySatisfactionBuildings.vue')['default']
     PlannerFactorySatisfactionItem: typeof import('./components/planner/PlannerFactorySatisfactionItem.vue')['default']
+    PlannerFactorySatisfactionItems: typeof import('./components/planner/PlannerFactorySatisfactionItems.vue')['default']
     PlannerGlobalActions: typeof import('./components/planner/PlannerGlobalActions.vue')['default']
     PlannerStatistics: typeof import('./components/planner/PlannerStatistics.vue')['default']
     RecipeItem: typeof import('./components/recipes/RecipeItem.vue')['default']
